# specs/003-user-auth/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Physical AI Textbook - Authentication API
  description: User authentication, registration, and rate limiting API for the Physical AI Humanoid Robotics Textbook
  version: 1.0.0
  contact:
    name: API Support

servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.physicalai.example.com/api/v1
    description: Production server

tags:
  - name: auth
    description: Authentication endpoints
  - name: user
    description: User profile management
  - name: rate-limit
    description: Rate limiting (anonymous users)

paths:
  /auth/register:
    post:
      tags: [auth]
      summary: Register a new user
      description: Create a new user account with email and password. Returns JWT tokens in HttpOnly cookies.
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          headers:
            Set-Cookie:
              description: HttpOnly cookies with access_token and refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error (invalid email, weak password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/login:
    post:
      tags: [auth]
      summary: Sign in with email and password
      description: Authenticate user and return JWT tokens in HttpOnly cookies.
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: HttpOnly cookies with access_token and refresh_token
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLockedResponse'

  /auth/logout:
    post:
      tags: [auth]
      summary: Sign out and terminate session
      description: Invalidate current session and clear authentication cookies.
      operationId: logoutUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Logout successful
          headers:
            Set-Cookie:
              description: Cleared cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: Use refresh token to get new access token. Called automatically by frontend on 401.
      operationId: refreshToken
      security:
        - refreshCookieAuth: []
      responses:
        '200':
          description: Token refreshed
          headers:
            Set-Cookie:
              description: New access_token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Token refreshed"
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/me:
    get:
      tags: [auth]
      summary: Get current user info
      description: Returns the currently authenticated user's basic info and profile.
      operationId: getCurrentUser
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithProfile'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /user/profile:
    post:
      tags: [user]
      summary: Create user profile (questionnaire)
      description: Save background questionnaire answers. Called after registration.
      operationId: createProfile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Profile already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags: [user]
      summary: Update user profile
      description: Update background questionnaire answers.
      operationId: updateProfile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileRequest'
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /rate-limit/status:
    get:
      tags: [rate-limit]
      summary: Get rate limit status for anonymous user
      description: Returns remaining requests and reset time for anonymous users.
      operationId: getRateLimitStatus
      parameters:
        - name: X-Fingerprint
          in: header
          description: Browser fingerprint for tracking
          schema:
            type: string
      responses:
        '200':
          description: Rate limit status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitStatus'

components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT access token in HttpOnly cookie
    refreshCookieAuth:
      type: apiKey
      in: cookie
      name: refresh_token
      description: JWT refresh token in HttpOnly cookie

  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
          example: "student@university.edu"
        password:
          type: string
          minLength: 8
          description: "Minimum 8 characters, at least 1 letter and 1 number"
          example: "SecurePass123"

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "student@university.edu"
        password:
          type: string
          example: "SecurePass123"

    ProfileRequest:
      type: object
      required: [programming_level, robotics_familiarity, hardware_experience, learning_goal]
      properties:
        programming_level:
          type: string
          enum: [none, beginner, intermediate, advanced]
          example: "intermediate"
        robotics_familiarity:
          type: string
          enum: [none, hobbyist, academic, professional]
          example: "hobbyist"
        hardware_experience:
          type: string
          enum: [none, arduino, embedded, industrial]
          example: "arduino"
        learning_goal:
          type: string
          enum: [career_change, academic, hobby, professional_dev]
          example: "academic"

    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserBasic'
        message:
          type: string
          example: "Registration successful"
        has_profile:
          type: boolean
          description: Whether user has completed questionnaire
          example: false

    UserBasic:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "student@university.edu"
        created_at:
          type: string
          format: date-time

    UserWithProfile:
      allOf:
        - $ref: '#/components/schemas/UserBasic'
        - type: object
          properties:
            profile:
              $ref: '#/components/schemas/ProfileResponse'

    ProfileResponse:
      type: object
      properties:
        programming_level:
          type: string
          example: "intermediate"
        robotics_familiarity:
          type: string
          example: "hobbyist"
        hardware_experience:
          type: string
          example: "arduino"
        learning_goal:
          type: string
          example: "academic"
        updated_at:
          type: string
          format: date-time

    RateLimitStatus:
      type: object
      properties:
        remaining:
          type: integer
          minimum: 0
          maximum: 5
          example: 3
        total:
          type: integer
          example: 5
        reset_at:
          type: string
          format: date-time
          description: When the rate limit window resets
        is_authenticated:
          type: boolean
          description: Always false for this endpoint (auth users have no limit)
          example: false

    AccountLockedResponse:
      type: object
      properties:
        error:
          type: string
          example: "Account temporarily locked"
        detail:
          type: string
          example: "Too many failed login attempts. Try again in 15 minutes."
        locked_until:
          type: string
          format: date-time
        code:
          type: integer
          example: 423

    ErrorResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          example: "Invalid credentials"
        detail:
          type: string
          example: "The email or password you entered is incorrect"
        code:
          type: integer
          example: 401